build: Add spark-4.0 profile
